+++ utils.py
@@ -14,6 +14,7 @@ edispkernel =irfs['edisp'].to_edisp_kernel(offset=1*u.deg)
 
 axis = np.log10(edispkernel.axes["energy_true"].center.value)
 axis = axis[18:227]
+axis = axis.astype(np.float128)
 energydisp = lambda log_energy_measured, log_energy_true: stats.norm(loc=log_energy_true, scale = 1e-3*(3-log_energy_true)+1e-3).pdf(log_energy_measured)
 
 # axis = np.linspace(-1.5,2.5,300)
@@ -32,7 +33,7 @@ def find_closest(arr, val):
 def make_gaussian(centre = 0.25, spread = 0.1, axis=axis):
     continuous_gaussian = stats.norm(loc=centre, scale=spread)
 
-    norm = np.sum(continuous_gaussian.pdf(axis))
+    norm = np.sum(continuous_gaussian.pdf(axis)).astype(np.float128)
     outputfunc = lambda x: continuous_gaussian.pdf(x)/norm
 
     return outputfunc
